Enable unit tests by default
When [tox] is invoked without specifying any environments, it runs
whatever is specified under `envlist` in its config. I had intended to
add the `unit` environment when  unit tests were added in 701188d but I
forgot. Fortunately this only applies to running the tests locally as
each environment is run separately in CI.

[tox]: https://tox.readthedocs.io