Base playlist names on common prefix of cue files
Currently, running this script on a folder that contains multiple `cue`
files will result in using the name of the first (alphabetically) `cue`
file as the basis for the name of the `m3u` file. This can lead to some
strange names like

    Some Game (Disc 1).m3u

This introduces two new functions that combine to reduce this down to a
name based on the common prefix shared by all `cue` files in the folder.

    Some Game.m3u

Closes #2